Problem: Annoying things: Moving instances to projects, and having to comment out already launched instances to prevent duplicates.
Solution: Wait until an instance has finished the 'build' phase before trying to launch the next one. This will prevent the 'Move to Project' dialog from popping up and breaking the tests. I also added a step to skip a scenario if an instance with the same name and provider already exists in the BDD project, making features simple and idempotent to run.